+++ packages/cli/src/buildsys/args.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Megaton contributors
 
 use cu::pre::*;
 #[derive(Debug, clap::Parser)]
