+++ packages/cli/src/buildsys/compile/source_type.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Megaton contributors
+
 use std::ffi::OsStr;
 use std::path::Path;
