+++ packages/cli/src/buildsys/lib_unpack.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Megaton contributors
+
 use std::path::Path;
 
 use cu::pre::*;
