+++ packages/cli/src/buildsys/miscfile.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Megaton contributors
 
 use std::path::Path;
