+++ packages/cli/src/cmds/cmd.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Megaton contributors
 
 use cu::pre::*;
